Move mutation query to LoginForm component and add redirecting user to redirectToPath after successful login
